+++ library/Library.Model/ReaderType.cs
@@ -11,7 +11,7 @@ namespace Library.Model
         public ReaderType() { }
         
         #region 读者类型表
-        public int rdType { get; set; }
+        public short rdType { get; set; }
         public string rdTypeName { get; set; }
         public int maxBorrowNum { get; set; }
         public int maxBorrowDay { get; set; }
